add more benchmarks

+++ BenchmarkCase.cpp
@@ -11,7 +11,7 @@ void BenchmarkCase::perform_case(double allowed_time) {
     ASSERT(allowed_time > 0);
 
     m_time_control.set_checkpoint();
-    uint64_t counter = 0;
+    long long int counter = 0;
     while (!m_time_control.is_elapsed(allowed_time)) {
         m_fct();
         counter++;
